% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptbuddy.R
\name{install_deps}
\alias{install_deps}
\title{Install dependencies detected from an R script or project}
\usage{
install_deps(path, cran_repo = "https://cloud.r-project.org", quiet = FALSE)
}
\description{
Install dependencies detected from an R script/project
}
\details{
Installs only missing packages from CRAN and Bioconductor.
}
