% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptbuddy.R
\name{scan_deps}
\alias{scan_deps}
\title{Scan dependencies from an R script or project folder}
\usage{
scan_deps(path)
}
\arguments{
\item{path}{Path to a \code{.R} file or a directory containing \code{.R} files.}
}
\value{
Character vector of detected package names.
}
\description{
ScriptBuddy finds packages used via \code{library()}, \code{require()}, \code{pkg::fun()}, and follows \code{source()} files recursively.
}
