% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptbuddy.R
\name{scriptbuddy_lock}
\alias{scriptbuddy_lock}
\title{Create a reproducible renv lockfile for a script/project}
\usage{
scriptbuddy_lock(
  path,
  project_dir = ".",
  cran_repo = "https://cloud.r-project.org"
)

scriptbuddy_lock(
  path,
  project_dir = ".",
  cran_repo = "https://cloud.r-project.org"
)
}
\arguments{
\item{path}{Script or folder to scan.}

\item{project_dir}{Directory where renv should live (default: current project).}

\item{cran_repo}{CRAN repo URL.}
}
\description{
This installs \code{renv} if needed, initializes renv (bare), and snapshots to \code{renv.lock}.
}
