% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptbuddy.R
\name{scriptbuddy_quiet}
\alias{scriptbuddy_quiet}
\title{Quiet wrapper for ScriptBuddy}
\usage{
scriptbuddy_quiet(
  path,
  install = FALSE,
  cran_repo = "https://cloud.r-project.org"
)
}
\arguments{
\item{path}{Script or folder to scan.}

\item{install}{If TRUE, installs missing packages.}

\item{cran_repo}{CRAN repo URL.}
}
\description{
Same as \code{scriptbuddy()} but suppresses verbose output.
}
