% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptbuddy.R
\name{scriptbuddy_report}
\alias{scriptbuddy_report}
\title{Pretty dependency report table}
\usage{
scriptbuddy_report(path, cran_repo = "https://cloud.r-project.org")
}
\arguments{
\item{path}{Script or folder to scan.}

\item{cran_repo}{CRAN repo URL.}
}
\description{
Pretty dependency report table
}
